+++ tests/test_e2e_ocr_conversion.py
@@ -1,3 +1,4 @@
+import sys
 from pathlib import Path
 from typing import List
 
@@ -6,8 +7,10 @@ from docling.datamodel.base_models import InputFormat
 from docling.datamodel.document import ConversionResult
 from docling.datamodel.pipeline_options import (
     EasyOcrOptions,
+    OcrMacOptions,
     OcrOptions,
     PdfPipelineOptions,
+    RapidOcrOptions,
     TesseractCliOcrOptions,
     TesseractOcrOptions,
 )
@@ -54,10 +57,17 @@ def test_e2e_conversions():
         EasyOcrOptions(),
         TesseractOcrOptions(),
         TesseractCliOcrOptions(),
+        RapidOcrOptions(),
         EasyOcrOptions(force_full_page_ocr=True),
         TesseractOcrOptions(force_full_page_ocr=True),
         TesseractCliOcrOptions(force_full_page_ocr=True),
+        RapidOcrOptions(force_full_page_ocr=True),
     ]
+
+    # only works on mac
+    if "darwin" == sys.platform:
+        engines.append(OcrMacOptions())
+        engines.append(OcrMacOptions(force_full_page_ocr=True))
 
     for ocr_options in engines:
         print(f"Converting with ocr_engine: {ocr_options.kind}")